Move comment-related functions to comments.py
